NetShaper: Fixed segmentation fault caused by limits not fitting in 32-bit integer (#5582)
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@41891 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/VBox/vmm/pdmnetshaper.h
@@ -99,7 +99,7 @@ VMMR3DECL(int) PDMR3NsDetach(PVM pVM, PPDMDRVINS pDrvIns, PPDMNSFILTER pFilter);
  * @param   pcszBwGroup           Name of the bandwidth group to attach to.
  * @param   cbTransferPerSecMax   Maximum number of bytes per second to be transmitted.
  */
-VMMR3DECL(int) PDMR3NsBwGroupSetLimit(PVM pVM, const char *pcszBwGroup, uint32_t cbTransferPerSecMax);
+VMMR3DECL(int) PDMR3NsBwGroupSetLimit(PVM pVM, const char *pcszBwGroup, uint64_t cbTransferPerSecMax);
 
 /** @} */
